+++ frontend/app/src/components/profitloss/ReportMissingAcquisitions.vue
@@ -181,9 +181,10 @@ export default defineComponent({
   }
 });
 </script>
+
 <style module lang="scss">
 .table {
-  max-height: calc(100vh - 350px);
+  max-height: calc(100vh - 250px);
   overflow: auto;
 }
 </style>