+++ frontend/app/src/components/status/sync/SyncIndicator.vue
@@ -27,86 +27,171 @@
           <v-icon v-else> mdi-content-save </v-icon>
         </menu-tooltip-button>
       </template>
-      <div :style="backgroundStyle">
+      <div>
         <div class="balance-saved-indicator__content">
-          <v-row v-if="premium">
-            <v-col>
-              <v-row no-gutters>
-                <v-col class="font-weight-medium">
-                  {{ $t('sync_indicator.last_data_upload') }}
-                </v-col>
-              </v-row>
-              <v-row class="text--secondary">
-                <v-col>
-                  <date-display
-                    v-if="lastDataUpload"
-                    :timestamp="lastDataUpload"
-                  />
-                  <span v-else>
-                    {{ $t('sync_indicator.never_saved') }}
-                  </span>
-                </v-col>
-              </v-row>
-              <v-row>
-                <v-col>
-                  <sync-buttons
-                    :pending="pending"
-                    :sync-action="syncAction"
-                    @action="showConfirmation($event)"
-                  />
-                </v-col>
-              </v-row>
-            </v-col>
-          </v-row>
-          <v-divider v-if="premium" class="mt-2" />
-          <v-row class="mt-2">
-            <v-col>
-              <v-row class="font-weight-medium" no-gutters>
-                <v-col>{{ $t('sync_indicator.snapshot_title') }}</v-col>
-              </v-row>
-              <v-row class="text--secondary">
-                <v-col>
-                  <date-display
-                    v-if="lastBalanceSave"
-                    :timestamp="lastBalanceSave"
-                  />
-                  <span v-else>
-                    {{ $t('sync_indicator.never_saved') }}
-                  </span>
-                </v-col>
-              </v-row>
-              <v-divider class="mt-2" />
-              <v-row align="center" class="mt-1">
-                <v-col>
-                  <v-btn color="primary" outlined @click="refreshAllAndSave()">
-                    <v-icon left>mdi-content-save</v-icon>
-                    {{ $t('sync_indicator.force_save') }}
-                  </v-btn>
-                  <v-tooltip right max-width="300px">
-                    <template #activator="{ on, attrs }">
-                      <div v-bind="attrs" v-on="on">
-                        <v-checkbox
-                          v-model="ignoreErrors"
-                          label="Ignore Errors"
-                        />
-                      </div>
-                    </template>
-                    <span>{{ $t('sync_indicator.ignore_errors') }}</span>
-                  </v-tooltip>
-                </v-col>
-                <v-col cols="auto">
-                  <v-tooltip bottom max-width="300px">
-                    <template #activator="{ on }">
-                      <v-icon class="ml-3" v-on="on">mdi-information</v-icon>
-                    </template>
-                    <div>
-                      {{ $t('sync_indicator.snapshot_tooltip') }}
+          <template v-if="premium">
+            <div class="font-weight-medium">
+              {{ $t('sync_indicator.last_data_upload') }}
+            </div>
+            <div class="py-2 text--secondary">
+              <date-display v-if="lastDataUpload" :timestamp="lastDataUpload" />
+              <span v-else>
+                {{ $t('sync_indicator.never_saved') }}
+              </span>
+            </div>
+            <div>
+              <sync-buttons
+                :pending="pending"
+                :sync-action="syncAction"
+                @action="showConfirmation($event)"
+              />
+            </div>
+            <v-divider class="my-4" />
+          </template>
+          <div>
+            <div class="font-weight-medium">
+              {{ $t('sync_indicator.snapshot_title') }}
+            </div>
+            <div class="pt-2 text--secondary">
+              <date-display
+                v-if="lastBalanceSave"
+                :timestamp="lastBalanceSave"
+              />
+              <span v-else>
+                {{ $t('sync_indicator.never_saved') }}
+              </span>
+            </div>
+            <v-divider class="my-4" />
+            <v-row>
+              <v-col>
+                <v-btn color="primary" outlined @click="refreshAllAndSave()">
+                  <v-icon left>mdi-content-save</v-icon>
+                  {{ $t('sync_indicator.force_save') }}
+                </v-btn>
+                <v-tooltip right max-width="300px">
+                  <template #activator="{ on, attrs }">
+                    <div v-bind="attrs" v-on="on">
+                      <v-checkbox
+                        v-model="ignoreErrors"
+                        hide-details
+                        label="Ignore Errors"
+                      />
                     </div>
-                  </v-tooltip>
-                </v-col>
-              </v-row>
-            </v-col>
-          </v-row>
+                  </template>
+                  <span>{{ $t('sync_indicator.ignore_errors') }}</span>
+                </v-tooltip>
+              </v-col>
+              <v-col cols="auto">
+                <v-tooltip bottom max-width="300px">
+                  <template #activator="{ on }">
+                    <v-icon class="ml-3" v-on="on">mdi-information</v-icon>
+                  </template>
+                  <div>
+                    {{ $t('sync_indicator.snapshot_tooltip') }}
+                  </div>
+                </v-tooltip>
+              </v-col>
+            </v-row>
+            <v-divider class="my-4" />
+            <div>
+              <div class="font-weight-medium">
+                {{ $t('sync_indicator.import_snapshot.title') }}
+              </div>
+              <div class="pt-4">
+                <v-dialog
+                  v-model="importSnapshotDialog"
+                  max-width="600"
+                  :persistent="
+                    !!balanceSnapshotFile || !!locationDataSnapshotFile
+                  "
+                >
+                  <template #activator="{ on }">
+                    <v-btn color="primary" outlined v-on="on">
+                      <v-icon left>mdi-import</v-icon>
+                      {{ $t('sync_indicator.import') }}
+                    </v-btn>
+                  </template>
+                  <card>
+                    <template #title>
+                      {{ $t('sync_indicator.import_snapshot.title') }}
+                    </template>
+                    <div class="pt-2">
+                      <v-row>
+                        <v-col>
+                          <div class="font-weight-bold">
+                            {{
+                              $t(
+                                'sync_indicator.import_snapshot.balance_snapshot_file'
+                              )
+                            }}
+                          </div>
+                          <div class="py-2">
+                            <file-upload
+                              ref="balanceSnapshotUploader"
+                              source="csv"
+                              @selected="balanceSnapshotFile = $event"
+                            />
+                          </div>
+                          <div class="text-caption">
+                            {{
+                              $t(
+                                'sync_indicator.import_snapshot.balance_snapshot_file_suggested'
+                              )
+                            }}
+                          </div>
+                        </v-col>
+                        <v-col>
+                          <div class="font-weight-bold">
+                            {{
+                              $t(
+                                'sync_indicator.import_snapshot.location_data_snapshot_file'
+                              )
+                            }}
+                          </div>
+                          <div class="py-2">
+                            <file-upload
+                              ref="locationDataSnapshotUploader"
+                              source="csv"
+                              @selected="locationDataSnapshotFile = $event"
+                            />
+                          </div>
+                          <div class="text-caption">
+                            {{
+                              $t(
+                                'sync_indicator.import_snapshot.location_data_snapshot_suggested'
+                              )
+                            }}
+                          </div>
+                        </v-col>
+                      </v-row>
+                    </div>
+                    <template #buttons>
+                      <v-spacer />
+                      <v-btn
+                        color="primary"
+                        text
+                        @click="importSnapshotDialog = false"
+                      >
+                        {{
+                          $t('sync_indicator.import_snapshot.actions.cancel')
+                        }}
+                      </v-btn>
+                      <v-btn
+                        color="primary"
+                        :disabled="!importFilesCompleted"
+                        :loading="importSnapshotLoading"
+                        @click="importSnapshot"
+                      >
+                        {{
+                          $t('sync_indicator.import_snapshot.actions.import')
+                        }}
+                      </v-btn>
+                    </template>
+                  </card>
+                </v-dialog>
+              </div>
+            </div>
+          </div>
         </div>
         <node-status v-if="xsOnly" :connected="nodeConnection" />
       </div>
@@ -120,7 +205,7 @@
       :primary-action="
         $tc('sync_indicator.upload_confirmation.action', textChoice)
       "
-      :secondary-action="$t('sync_indicator.upload_confirmation.cancel')"
+      :secondary-action="$tc('sync_indicator.upload_confirmation.cancel')"
       @cancel="cancel"
       @confirm="performSync"
     >
@@ -139,22 +224,30 @@
   </fragment>
 </template>
 <script lang="ts">
-import { Component, Mixins } from 'vue-property-decorator';
-import { mapActions, mapState } from 'vuex';
+import { computed, defineComponent, ref } from '@vue/composition-api';
+import { get, set } from '@vueuse/core';
 import ConfirmDialog from '@/components/dialogs/ConfirmDialog.vue';
 import DateDisplay from '@/components/display/DateDisplay.vue';
 import Fragment from '@/components/helper/Fragment';
 import MenuTooltipButton from '@/components/helper/MenuTooltipButton.vue';
+import FileUpload from '@/components/import/FileUpload.vue';
 import NodeStatus from '@/components/status/NodeStatus.vue';
 import SyncButtons from '@/components/status/sync/SyncButtons.vue';
-import PremiumMixin from '@/mixins/premium-mixin';
-import ThemeMixin from '@/mixins/theme-mixin';
+import { setupGeneralBalances } from '@/composables/balances';
+import { setupThemeCheck } from '@/composables/common';
+import { getPremium, setupSession } from '@/composables/session';
+import { interop } from '@/electron-interop';
+import i18n from '@/i18n';
+import { api } from '@/services/rotkehlchen-api';
 import { SYNC_DOWNLOAD, SYNC_UPLOAD, SyncAction } from '@/services/types-api';
 import { AllBalancePayload } from '@/store/balances/types';
+import { showError, showMessage } from '@/store/utils';
 import { Writeable } from '@/types';
 
-@Component({
+export default defineComponent({
+  name: 'SyncIndicator',
   components: {
+    FileUpload,
     NodeStatus,
     Fragment,
     ConfirmDialog,
@@ -162,79 +255,164 @@ import { Writeable } from '@/types';
     DateDisplay,
     MenuTooltipButton
   },
-  computed: {
-    ...mapState('session', [
-      'lastBalanceSave',
-      'lastDataUpload',
-      'nodeConnection'
-    ])
-  },
-  methods: {
-    ...mapActions('balances', ['fetchBalances']),
-    ...mapActions('session', ['forceSync'])
-  }
-})
-export default class SyncIndicator extends Mixins(PremiumMixin, ThemeMixin) {
-  pending: boolean = false;
-  confirmChecked: boolean = false;
-  ignoreErrors: boolean = false;
-  visible: boolean = false;
-  syncAction: SyncAction = SYNC_UPLOAD;
-  lastBalanceSave!: number;
-  lastDataUpload!: string;
-  forceSync!: (action: SyncAction) => Promise<void>;
-  fetchBalances!: (payload: Partial<AllBalancePayload>) => Promise<void>;
-  displayConfirmation: boolean = false;
+  setup() {
+    const { lastBalanceSave, lastDataUpload, nodeConnection, forceSync } =
+      setupSession();
 
-  get xsOnly(): boolean {
-    return this.$vuetify.breakpoint.xsOnly;
-  }
+    const { fetchBalances } = setupGeneralBalances();
+    const { currentBreakpoint } = setupThemeCheck();
+    const premium = getPremium();
 
-  get isDownload(): boolean {
-    return this.syncAction === SYNC_DOWNLOAD;
-  }
+    const pending = ref<boolean>(false);
+    const confirmChecked = ref<boolean>(false);
+    const ignoreErrors = ref<boolean>(false);
+    const visible = ref<boolean>(false);
+    const syncAction = ref<SyncAction>(SYNC_UPLOAD);
+    const displayConfirmation = ref<boolean>(false);
+    const balanceSnapshotUploader = ref<any>(null);
+    const balanceSnapshotFile = ref<File | null>(null);
+    const locationDataSnapshotUploader = ref<any>(null);
+    const locationDataSnapshotFile = ref<File | null>(null);
+    const importSnapshotLoading = ref<boolean>(false);
+    const importSnapshotDialog = ref<boolean>(false);
 
-  get textChoice(): number {
-    return this.syncAction === SYNC_UPLOAD ? 1 : 2;
-  }
+    const xsOnly = computed(() => get(currentBreakpoint).xsOnly);
 
-  get message(): string {
-    return this.syncAction === SYNC_UPLOAD
-      ? this.$tc('sync_indicator.upload_confirmation.message_upload')
-      : this.$tc('sync_indicator.upload_confirmation.message_download');
-  }
+    const isDownload = computed<boolean>(
+      () => get(syncAction) === SYNC_DOWNLOAD
+    );
+    const textChoice = computed<number>(() =>
+      get(syncAction) === SYNC_UPLOAD ? 1 : 2
+    );
+    const message = computed<string>(() => {
+      return get(syncAction) === SYNC_UPLOAD
+        ? i18n.t('sync_indicator.upload_confirmation.message_upload').toString()
+        : i18n
+            .t('sync_indicator.upload_confirmation.message_download')
+            .toString();
+    });
 
-  async refreshAllAndSave() {
-    this.visible = false;
-    const payload: Writeable<Partial<AllBalancePayload>> = {
-      ignoreCache: true,
-      saveData: true
+    const refreshAllAndSave = async () => {
+      set(visible, false);
+      const payload: Writeable<Partial<AllBalancePayload>> = {
+        ignoreCache: true,
+        saveData: true
+      };
+      if (get(ignoreErrors)) {
+        payload.ignoreErrors = true;
+      }
+      await fetchBalances(payload);
     };
 
-    if (this.ignoreErrors) {
-      payload.ignoreErrors = true;
-    }
-    await this.fetchBalances(payload);
-  }
+    const showConfirmation = (action: SyncAction) => {
+      set(visible, false);
+      set(syncAction, action);
+      set(displayConfirmation, true);
+    };
 
-  showConfirmation(action: SyncAction) {
-    this.visible = false;
-    this.syncAction = action;
-    this.displayConfirmation = true;
-  }
+    const performSync = async () => {
+      set(pending, true);
+      await forceSync(get(syncAction));
+      set(pending, false);
+      set(displayConfirmation, false);
+      set(confirmChecked, false);
+    };
 
-  performSync() {
-    this.pending = true;
-    this.forceSync(this.syncAction).then(() => (this.pending = false));
-    this.displayConfirmation = false;
-    this.confirmChecked = false;
-  }
+    const cancel = () => {
+      set(displayConfirmation, false);
+      set(confirmChecked, false);
+    };
 
-  cancel() {
-    this.displayConfirmation = false;
-    this.confirmChecked = false;
+    const importFilesCompleted = computed<boolean>(
+      () => !!get(balanceSnapshotFile) && !!get(locationDataSnapshotFile)
+    );
+
+    const { logout } = setupSession();
+
+    const importSnapshot = async () => {
+      if (!get(importFilesCompleted)) return;
+      set(importSnapshotLoading, true);
+
+      let success = false;
+      let message = '';
+      try {
+        if (interop.appSession) {
+          await api.importBalancesSnapshot(
+            get(balanceSnapshotFile)!.path,
+            get(locationDataSnapshotFile)!.path
+          );
+        } else {
+          await api.uploadBalancesSnapshot(
+            get(balanceSnapshotFile)!,
+            get(locationDataSnapshotFile)!
+          );
+        }
+        success = true;
+      } catch (e: any) {
+        message = e.message;
+      }
+
+      if (!success) {
+        showError(
+          i18n
+            .t('sync_indicator.import_snapshot.messages.failed_description', {
+              message
+            })
+            .toString(),
+          i18n.t('sync_indicator.import_snapshot.messages.title').toString()
+        );
+      } else {
+        showMessage(
+          i18n
+            .t('sync_indicator.import_snapshot.messages.success_description', {
+              message
+            })
+            .toString(),
+          i18n.t('sync_indicator.import_snapshot.messages.title').toString()
+        );
+
+        setTimeout(() => {
+          logout();
+        }, 3000);
+      }
+
+      set(importSnapshotLoading, false);
+      get(balanceSnapshotUploader)?.removeFile();
+      get(locationDataSnapshotUploader)?.removeFile();
+      set(balanceSnapshotFile, null);
+      set(locationDataSnapshotFile, null);
+    };
+
+    return {
+      xsOnly,
+      premium,
+      lastBalanceSave,
+      lastDataUpload,
+      nodeConnection,
+      pending,
+      confirmChecked,
+      ignoreErrors,
+      visible,
+      syncAction,
+      displayConfirmation,
+      balanceSnapshotUploader,
+      locationDataSnapshotUploader,
+      balanceSnapshotFile,
+      locationDataSnapshotFile,
+      importSnapshotLoading,
+      importSnapshotDialog,
+      importFilesCompleted,
+      textChoice,
+      message,
+      isDownload,
+      cancel,
+      performSync,
+      showConfirmation,
+      refreshAllAndSave,
+      importSnapshot
+    };
   }
-}
+});
 </script>
 
 <style lang="scss" scoped>